Test + additional fixes
